refactor: Extract search logic from SearchBar to useSearch hook
